IMP - Refactored components
