+++ src/components/elements/gameData/skillAtk/in/sectionEx.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import {useI18n} from '../../../../../i18n/hook';
-import {InlineCheck} from '../../elements/inlineCheck';
+import {InlineCheck} from '../../../common/check/inlineCheck';
 import {SectionTitle} from '../../elements/title';
 import {SectionProps} from './props';
