+++ src/components/elements/gameData/skillAtk/in/sectionFilter.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import {Form} from 'react-bootstrap';
 
 import {useI18n} from '../../../../../i18n/hook';
-import {EnumChecksBox} from '../../../common/enum/enumChecksBox';
-import {InlineCheck} from '../../elements/inlineCheck';
+import {EnumChecksBox} from '../../../common/check/enumChecksBox';
+import {InlineCheck} from '../../../common/check/inlineCheck';
 import {SectionSubTitle} from '../../elements/subTitle';
 import {SectionTitle} from '../../elements/title';
 import {
