+++ src/components/elements/gameData/skillAtk/in/sectionOther.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import {useI18n} from '../../../../../i18n/hook';
-import {NumericInput} from '../../elements/numInput';
+import {NumericInput} from '../../../common/input/numeric';
 import {SectionTitle} from '../../elements/title';
 import {SectionProps} from './props';
